backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.26.3
anthropic==0.18.1
boto3==1.34.34
python-dotenv==1.0.1
filterpy==1.4.5
